Separated Items login from main file
